refactor(ulw-loop): single-pass defect aggregation and artifact-primitive extraction
